Move path constants from mslib.msui to mslib.utils
