+++ mslib/msui/mscolab.py
@@ -63,7 +63,7 @@ from mslib.msui.qt5 import ui_mscolab_help_dialog as msc_help_dialog
 from mslib.msui.qt5 import ui_mscolab_add_operation_dialog as msc_add_operation_ui
 from mslib.msui.qt5 import ui_mscolab_merge_waypoints_dialog as merge_wp_ui
 from mslib.msui.qt5 import ui_mscolab_profile_dialog as ui_profile
-from mslib.msui import constants
+from mslib.utils import constants
 from mslib.utils.config import config_loader
