+++ mslib/utils/migration/config_before_eight.py
@@ -37,7 +37,7 @@ import os
 import tempfile
 
 from mslib.utils import FatalUserError
-from mslib.msui import constants
+from mslib.utils import constants
 from mslib.support.qt_json_view.datatypes import match_type, UrlType, StrType
