+++ mslib/utils/migration/update_json_file_to_version_eight.py
@@ -39,7 +39,7 @@ from mslib.utils.migration.config_before_eight import read_config_file as read_c
 from mslib.utils.migration.config_before_eight import config_loader as config_loader_before_eight
 from mslib.utils.config import modify_config_file
 from mslib.utils.config import read_config_file, config_loader
-from mslib.msui.constants import MSUI_SETTINGS
+from mslib.utils.constants import MSUI_SETTINGS
 
 
 class JsonConversion:
