+++ tests/_test_msui/test_sideview.py
@@ -155,7 +155,7 @@ class Test_MSSSideViewWindow:
         In tutorial mode the displayed flight track is silently written to a
         fixed FTML, and the file is rewritten when the model changes.
         """
-        from mslib.msui import constants
+        from mslib.utils import constants
         ftml = os.path.join(constants.MSUI_CONFIG_PATH, "tutorial_flighttrack.ftml")
         if os.path.exists(ftml):
             os.remove(ftml)
@@ -182,7 +182,7 @@ class Test_MSSSideViewWindow:
         so the creator calls enable_tutorial_flighttrack_save to write the initial
         FTML and start tracking (mscolab).
         """
-        from mslib.msui import constants
+        from mslib.utils import constants
         ftml = os.path.join(constants.MSUI_CONFIG_PATH, "tutorial_flighttrack.ftml")
         if os.path.exists(ftml):
             os.remove(ftml)
