+++ tests/_test_utils/test_migration.py
@@ -30,7 +30,7 @@ from packaging import version
 
 from mslib.utils import auth
 from mslib import __version__
-from mslib.msui.constants import MSUI_SETTINGS
+from mslib.utils.constants import MSUI_SETTINGS
 from tests.utils import create_msui_settings_file
