+++ tutorials/utils/picture.py
@@ -28,7 +28,7 @@ import time
 from pathlib import Path
 from slugify import slugify
 from PIL import Image
-from mslib.msui.constants import MSUI_CONFIG_PATH
+from mslib.utils.constants import MSUI_CONFIG_PATH
 
 
 def picture(name, bounding_box=None):
